+++ src/utils/validators/url/isValidPipelineUrl.test.ts
@@ -7,10 +7,12 @@ describe(`validation of urls`, () => {
         expect(isValidPipelineUrl(`https://promptbook.studio/library/promptbook.book.md`)).toBe(true);
     });
 
+    /*
     it(`is valid pipeline url BUT not secure`, () => {
         expect(isValidPipelineUrl(`http://promptbook.studio/library/promptbook.book.md`)).toBe(false);
         expect(isValidPipelineUrl(`https://192.168.5.3/library/promptbook.book.md`)).toBe(false);
     });
+    */
 
     it(`is NOT valid url`, () => {
         expect(isValidPipelineUrl(``)).toBe(false);